aisupporter/requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
langchain>=0.0.344
langchain-openai>=0.0.1
langgraph>=0.0.20
duckdb==0.8.1
pandas==2.0.1
numpy>=1.26.0
plotly>=5.18.0
pydantic>=2.4.2
python-dotenv==1.0.0
pytest>=7.4.3
httpx>=0.25.0
opencv-python>=4.8.1.78
pillow>=10.1.0
python-multipart>=0.0.6
jinja2>=3.1.2
aiofiles>=23.2.1
pyarrow==14.0.1
torch>=2.1.0
transformers>=4.35.0
openai>=1.1.1
Flask==2.3.3
Flask-WTF==1.2.1
WTForms==3.0.1
pymysql==1.1.0
flask-sqlalchemy==3
flask-login==0.6.2
email-validator==2.1.0.post1
flask-migrate==4.0.5
werkzeug==2.3.7
itsdangerous==2.1.2
gunicorn==20.1.0
pytest>=7.3.1
fastparquet==2023.10.1
sqlalchemy==2.0.22
sqlalchemy-utils==0.41.1
alembic==1.12.0
xlsxwriter==3.1.2
matplotlib==3.8.0 